fix(web): split setup completion flag for Edge middleware
setupComplete.ts pulled node:fs/path into middleware via setupMode.
Extract edge-safe SETUP_COMPLETE_ENV + isSetupCompleteFlagSet into
setupCompleteFlag.ts; keep marker-file logic Node-only.

Fixes Turbopack build failures on PR #250 (Docker, perimeter smoke,
Playwright smoke, Vercel).
